practice-1/pyproject.toml
[tool.poetry]
name = "practice-1"
version = "0.1.0"
description = ""
authors = ["Rustam Akimov <akirusprod@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.32.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
